+++ Assets/Scripts/Enemy/WaveManager.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
+using Control;
 using UnityEngine;
 
 public class WaveManager : MonoBehaviour
